+++ src/commonMain/kotlin/csense/kotlin/datastructures/caching/ValueWithCacheComputation.kt
@@ -4,26 +4,35 @@ public class ValueWithCacheComputation<Value, CachedType>(
     private val initialValue: Value,
     private val calculateCache: (Value) -> CachedType
 ) {
-
-    private var _cachedValue: CachedType? = null
-    public val cachedValue: CachedType
-        get() = _cachedValue ?: calculateCache(value).also {
-            _cachedValue = it
-        }
-
-    public var value: Value = initialValue
-        set(newValue) {
-            if (newValue != value) {
-                field = newValue
-                invalidateCache()
-            }
-        }
-
-    private fun invalidateCache() {
-        _cachedValue = null
+    private var _cachedValue: CachedComputableValue<CachedType> = CachedComputableValue {
+        calculateCache(value)
     }
 
+    /**
+     * The cached value (based on the value)
+     * If this has not been computed it will be on retrieval
+     */
+    public val cachedValue: CachedType by _cachedValue
+
+    /**
+     * The stored value
+     * getting has no side effect
+     * setting has the side effect of invaliding the current cached value
+     */
+    public var value: Value = initialValue
+        set(newValue) {
+            if (newValue == value) {
+                return
+            }
+            field = newValue
+            _cachedValue.invalidate()
+        }
+
+    /**
+     * Resets the value to the initial value (this also invalidates the cached value)
+     */
     public fun resetToInitial() {
         value = initialValue
     }
 }
+